+++ Runtime/Scripts/Base/General.cs
@@ -17,6 +17,7 @@ namespace Kokowolo.Utilities
 {
     public static class General
     {
+        [System.Obsolete("please write an encapsulated property delcaration instead")]
         public static T CacheGetComponent<T>(this MonoBehaviour monoBehaviour, ref T component) where T : Component
         {
             if (!component) component = monoBehaviour.GetComponent<T>();